feat: lift provider and update types
