+++ apps/web/src/pages/search.tsx
@@ -3,8 +3,7 @@ import { useState } from 'react';
 import { GetServerSideProps } from 'next';
 import Head from 'next/head';
 import { Flex, Image, Text, Textarea } from '@chakra-ui/react';
-import { TweetData } from '../types/Tweet';
-import { Timeline } from '../components/Timeline';
+import { Timeline, TweetData } from '../components/Timeline';
 import { useDebouncedCallback } from 'use-debounce';
 import Router from 'next/router';
